refactor: import through the modules the barrels re-exported
unicorn/no-barrel-files rejects a module that only re-exports. Ten of these
are removed and their importers point at the real module instead; six survive
where the file has content of its own, and eslint.config.js lists them.

The four vi.mock factories that named a removed barrel are repointed too: a
mock path is invisible to an import search and only vue-tsc caught them.
