--- frontend/app/electron/main/ipc-handlers/index.ts
@@ -1,13 +1,0 @@
-export { BackendHandlers } from './backend-handlers';
-
-export { OAuthHandlers } from './oauth-handlers';
-
-export { SecurityHandlers } from './security-handlers';
-
-export { SystemHandlers } from './system-handlers';
-
-export { UpdateHandlers } from './update-handlers';
-
-export { WalletBridgeIpcHandlers } from './wallet-bridge-ipc-handlers';
-
-export { WalletImportHandlers } from './wallet-import-handlers';
