--- frontend/app/src/modules/accounts/table/index.ts
@@ -1,1 +1,0 @@
-export { default as AccountBalancesTable } from './AccountBalancesTable.vue';
