--- frontend/app/src/modules/assets/detection/index.ts
@@ -1,3 +1,0 @@
-export { NewDetectedToken } from './types';
-
-export { useNewlyDetectedTokens } from './use-newly-detected-tokens';
