--- frontend/app/src/modules/core/api/index.ts
@@ -1,5 +1,0 @@
-export { RequestCancelledError } from './request-queue/errors';
-
-export { api } from './rotki-api';
-
-export { snakeCaseTransformer } from './transformers';
