+++ frontend/app/src/modules/history/api/events/use-asset-movement-matching-api.ts
@@ -1,5 +1,5 @@
-import { snakeCaseTransformer } from '@/modules/core/api';
 import { api } from '@/modules/core/api/rotki-api';
+import { snakeCaseTransformer } from '@/modules/core/api/transformers';
 import { type PendingTask, PendingTaskSchema } from '@/modules/core/tasks/types';
 import { useTaskApi } from '@/modules/core/tasks/use-task-api';
