--- frontend/app/src/modules/history/balances/components.ts
@@ -1,3 +1,0 @@
-export { default as AccountingOverlayToggle } from '@/modules/history/balances/AccountingOverlayToggle.vue';
-
-export { default as BalanceDivergenceToggle } from '@/modules/history/balances/BalanceDivergenceToggle.vue';
