--- frontend/app/tests/unit/setup-files/handlers/index.ts
@@ -1,31 +1,0 @@
-export * from './all-evm-chains';
-
-export * from './assets';
-
-export * from './binance-savings';
-
-export * from './history-event-counterparties';
-
-export * from './history-event-products';
-
-export * from './history-events';
-
-export * from './history-type-mappings';
-
-export * from './info';
-
-export * from './nfts';
-
-export * from './premium';
-
-export * from './queried-addresses';
-
-export * from './settings';
-
-export * from './skipped-external-events';
-
-export * from './staking';
-
-export * from './supported-chains';
-
-export * from './task-scheduler';
