--- frontend/common/src/settings/index.ts
@@ -1,5 +1,0 @@
-export * from './frontend';
-
-export * from './graphs';
-
-export * from './themes';
