--- frontend/common/src/staking/index.ts
@@ -1,1 +1,0 @@
-export * from './eth2';
